refactor(admin): move useDashboardSummary hook and test out of components/admin
Addresses PR #119 review comments r3107522748 and r3107521277 from @Jorenve.

Hook moved to frontend/src/components/admin/hooks/useDashboardSummary.ts.
Placing it in a hooks/ subdirectory keeps admin-specific hooks co-located with
the admin area (consistent with useTheme.ts, useLocale.ts, useFocusTrap.ts all
living under components/admin/) while satisfying the reviewer's request to move
it out of the components/admin/ root.

Test moved to frontend/src/test/components/admin/useDashboardSummary.test.tsx,
mirroring the existing test/components/admin/ convention used by AdminLayout.test.tsx.

Updated import paths in:
- frontend/src/components/admin/hooks/useDashboardSummary.ts (internal api/client import)
- frontend/src/pages/admin/DashboardPage.tsx (consumer import)
- frontend/src/test/pages/admin/DashboardPage.test.tsx (vi.mock path)
- frontend/src/test/components/admin/useDashboardSummary.test.tsx (hook + api imports)
