+++ frontend/src/components/admin/hooks/useDashboardSummary.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { api } from '../../api/client'
+import { api } from '../../../api/client'
 
 type Delta = {
     changePct: number | null
